+++ CHANGELOG.rst
@@ -19,6 +19,12 @@ Added
 * Input type for **start_tracker_thread** method for param **state_function** is changed from **Callable** to **str**
 * **timeout_decorator** and **error_propagation_decorator** added for implementing timeout and error propagation functionalities respectively
 
+[0.15.0]
+************
+* Updated Scan command interface to include scan_id as argument
+* EndScan command has been added in HelperDishLNDevice
+* **scanID** attribute has been introduced in HelperDishDevice
+
 
 Fixed
 -----